pyproject.toml
[project]
name = "cv-search-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcin Zielinski",email = "marcin.a.zielinski@accenture.com"}
]
requires-python = ">=3.13, <3.14"
dependencies = [
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "chromadb (>=1.5.0,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "ollama (>=0.6.1,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "rag_chatbot", from = "src" }]
